requirements.txt
openai-whisper
pydub
jiwer
numpy
pandas
matplotlib
seaborn
scipy